chore: move the template folder directory
