+++ apps/editor/src/new/vdom/template.ts
@@ -1,8 +1,7 @@
 import html from 'htm/mini';
-import isFunction from 'tui-code-snippet/type/isFunction';
 import isString from 'tui-code-snippet/type/isString';
 import { ComponentClass } from '@t/ui';
-import { VNode } from './vdom/vnode';
+import { VNode } from './vnode';
 
 function createTextNode(text: string) {
   return new VNode('TEXT_NODE', { nodeValue: text }, []);
@@ -23,10 +22,6 @@ function h(type: string | ComponentClass, props: Record<string, any> = {}, ...ch
 
   const vnode = new VNode(type, props, flatted);
 
-  if (!isFunction(type)) {
-    vnode.ref = props.ref;
-    delete props.ref;
-  }
   return vnode;
 }
